requirements.txt
python-telegram-bot[job-queue]
beautifulsoup4
requests